+++ packages/site/src/componentList.ts
@@ -65,6 +65,16 @@ export const componentList = [
     imageURL: "/Countdown.png",
   },
   {
+    title: "DataDump",
+    to: "/components/DataDump",
+    imageURL: "/DataDump.png",
+  },
+  {
+    title: "DataList",
+    to: "/components/DataList",
+    imageURL: "/DataList.png",
+  },
+  {
     title: "Disclosure",
     to: "/components/Disclosure",
     imageURL: "/Disclosure.png",
@@ -81,6 +91,11 @@ export const componentList = [
     to: "/components/Emphasis",
     imageURL: "/Emphasis.png",
     additionalMatches: ["Highlight", "Strong", "Bold", "Italic"],
+  },
+   {
+    title: "FormatDate",
+    to: "/components/FormatDate",
+    imageURL: "/FormatDate.png",
   },
   {
     title: "Glimmer",