+++ packages/site/src/content/index.ts
@@ -9,8 +9,11 @@ import CardContent from "./Card";
 import ChipContent from "./Chip";
 import ChipsContent from "./Chips";
 import CountdownContent from "./Countdown";
+import DataDumpContent from "./DataDump";
+import DataListContent from "./DataList";
 import DisclosureContent from "./Disclosure";
 import EmphasisContent from "./Emphasis";
+import FormatDateContent from "./FormatDate";
 import GlimmerContent from "./Glimmer";
 import HeadingContent from "./Heading";
 import IconContent from "./Icon";
@@ -59,11 +62,20 @@ export const SiteContent: Record<string, ContentExport> = {
   Countdown: {
     ...CountdownContent,
   },
+  DataDump: {
+    ...DataDumpContent,
+  },
+  DataList: {
+    ...DataListContent,
+  },
   Disclosure: {
     ...DisclosureContent,
   },
   Emphasis: {
     ...EmphasisContent,
+  },
+  FormatDate: {
+    ...FormatDateContent,
   },
   Glimmer: {
     ...GlimmerContent,